+++ packages/db-commons/index.d.ts
@@ -2,7 +2,8 @@ export * from './index.cjs';
 
 export type RunQuery<T extends Record<string, unknown>> = (
 	queryString: string,
-	database: T
+	database: T,
+	batchSize: number
 ) => Promise<QueryResult>;
 
 export type EvidenceColumnType = number | boolean | string | Date;
@@ -10,4 +11,4 @@ export type EvidenceColumnType = number | boolean | string | Date;
 export type GetRunner<T extends Record<string, unknown>> = (
 	opts: T,
 	directory: string
-) => (queryContent: string, queryPath: string) => Promise<QueryResult>;
+) => (queryContent: string, queryPath: string, batchSize: number) => Promise<QueryResult>;