+++ sites/example-project/src/pages/+layout.server.js
@@ -1,13 +1,14 @@
 import { getQueries } from '@evidence-dev/db-orchestrator';
 import md5 from 'blueimp-md5';
 import { GET } from './api/customFormattingSettings.json/+server.js';
+import { getStatusAndExtractQueries } from './extractQueries.server.js';
 export const prerender = true;
 export const trailingSlash = 'always';
 
 const system_routes = ['/settings', '/explore'];
 
 /** @satisfies {import("./$types").LayoutServerLoad} */
-export async function load({ fetch, route, params }) {
+export async function load({ route, params }) {
 	const isUserPage =
 		route.id && system_routes.every((system_route) => !route.id.startsWith(system_route));
 
@@ -20,9 +21,8 @@ export async function load({ fetch, route, params }) {
 	);
 
 	if (isUserPage) {
-		// ensure that queries have been extracted before initiating the load process
-		const statusEndpoint = `/api/status${route.id}`.replace(/\/$/, '');
-		await fetch(statusEndpoint);
+		// todo: remove this
+		await getStatusAndExtractQueries(route.id);
 	}
 
 	const customFormattingSettingsRes = await GET();